requirements.txt
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.2
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
requests==2.31.0
SQLAlchemy==2.0.25
typing_extensions==4.9.0
urllib3==1.26.18
Werkzeug==3.0.1
zipp==3.17.0
python-dotenv~=1.0.1
psycopg2==2.9.9
